wasm: moving a utility to wasm/common
