+++ source/extensions/common/wasm/wasm.h
@@ -21,6 +21,7 @@
 #include "source/common/version/version.h"
 #include "source/extensions/common/wasm/context.h"
 #include "source/extensions/common/wasm/plugin.h"
+#include "source/extensions/common/wasm/remote_async_datasource.h"
 #include "source/extensions/common/wasm/stats_handler.h"
 #include "source/extensions/common/wasm/wasm_vm.h"
 
@@ -170,8 +171,7 @@ bool createWasm(const PluginSharedPtr& plugin, const Stats::ScopeSharedPtr& scop
                 Upstream::ClusterManager& cluster_manager, Init::Manager& init_manager,
                 Event::Dispatcher& dispatcher, Api::Api& api,
                 Envoy::Server::ServerLifecycleNotifier& lifecycle_notifier,
-                Config::DataSource::RemoteAsyncDataProviderPtr& remote_data_provider,
-                CreateWasmCallback&& callback,
+                RemoteAsyncDataProviderPtr& remote_data_provider, CreateWasmCallback&& callback,
                 CreateContextFn create_root_context_for_testing = nullptr);
 
 PluginHandleSharedPtr
