+++ source/extensions/filters/http/wasm/wasm_filter.h
@@ -8,6 +8,7 @@
 #include "envoy/upstream/cluster_manager.h"
 
 #include "source/extensions/common/wasm/plugin.h"
+#include "source/extensions/common/wasm/remote_async_datasource.h"
 #include "source/extensions/common/wasm/wasm.h"
 
 namespace Envoy {
@@ -51,7 +52,7 @@ public:
 
 private:
   ThreadLocal::TypedSlotPtr<PluginHandleSharedPtrThreadLocal> tls_slot_;
-  Config::DataSource::RemoteAsyncDataProviderPtr remote_data_provider_;
+  RemoteAsyncDataProviderPtr remote_data_provider_;
 };
 
 using FilterConfigSharedPtr = std::shared_ptr<FilterConfig>;
